CreateDiscussionModal & fixes
