+++ src/App.tsx
@@ -20,7 +20,6 @@ import LoginPage from "./pages/identity/LoginPage";
 import RegisterPage from "./pages/identity/RegisterPage";
 import Agenda from "./pages/Agenda";
 import ClubPage from "./pages/clubs/ClubPage";
-import AddDiscussion from "./pages/discussion/CreateDiscussion";
 import LiveDiscussion from "./pages/discussion/LiveDiscussion";
 import CreateResource from "./pages/resources/CreateResource";
 import StartPage from "./pages/identity/StartPage";
@@ -95,7 +94,6 @@ return <IonApp>
               <PrivateRoute path="/clubs/:bookClubId/edit" component={EditClubPage} exact/>
               <PrivateRoute path="/clubs/:bookClubId/resources/add" component={CreateResource} exact/>
               <PrivateRoute path="/clubs/:bookClubId/resources/:resourceId/edit" component={EditResource} exact/>
-              <PrivateRoute path="/clubs/:bookClubId/discussions/add" component={AddDiscussion} exact/>
               <PrivateRoute path="/clubs/:bookClubId/discussions/:discussionId/comments" component={Comments} exact/>
               <PrivateRoute path="/clubs/:bookClubId/discussions/:discussionId/edit" component={EditDiscussion} exact/>
               <PrivateRoute path="/clubs/:bookClubId/discussions/:discussionId/comments/:commentId/edit" component={EditComment} exact/>
