--- src/pages/discussion/CreateDiscussion.css
@@ -1,11 +1,0 @@
-.testclass {
-  margin-left: 100px;
-}
-
-.divider {
-  height: 30px;
-}
-
-.box {
-  margin-left: 30px;
-}
